[ADDコードをライブラリクレートに分割する
